+++ flirror/exceptions.py
@@ -14,3 +14,9 @@ class GoogleOAuthError(Exception):
     """Exception if the Google OAuth authencitaion failed."""
 
     pass
+
+
+class ModuleDataException(Exception):
+    """Exeption if the data for a module could not be retrieved."""
+
+    pass
